slight modifications

+++ data_API/data_downloader_chestXray.py
@@ -21,8 +21,9 @@ links = [
 
 for idx, link in enumerate(links):
     fn = '../data/images_%02d.tar.gz' % (idx+1)
-    print('downloading'+fn+'...')
-    urllib.request.urlretrieve(link, fn)  # download the zip file
+    if not os.path.exists(fn) :
+        print('downloading'+fn+'...')
+        urllib.request.urlretrieve(link, fn)  # download the zip file
 
 print("Download complete. Please check the checksums")
